+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/user/updateuser/UpdateResponseDto.kt
@@ -1,0 +1,13 @@
+package com.app.arcabyolimpo.data.remote.dto.user.updateuser
+
+import com.google.gson.annotations.SerializedName
+
+data class UpdateResponseDto(
+    @SerializedName("status")
+    val status: Boolean,
+
+    @SerializedName("message")
+    val message: String
+)
+
+